refactor(order): migrate discount type to Integer (0 - 100)

+++ src/main/java/com/progra3/cafeteria_api/model/dto/OrderResponseDTO.java
@@ -16,7 +16,7 @@ public record OrderResponseDTO (
     LocalDate date,
     LocalTime time,
     Integer peopleCount,
-    Double discount,
+    Integer discount,
     String status,
     Double subtotal,
     Double total
